requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
numpy==2.2.3
rdflib==7.1.3
paho-mqtt==2.1.0
pydantic==2.11.0a2
pydantic[email]
python-jose==3.3.0
cryptography==44.0.1
passlib==1.7.4
email-validator==2.0.0
python-multipart==0.0.20